+++ src/FieldLayout/Contracts/FieldLayoutFormInputProviderInterface.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace CraftCms\Cms\FieldLayout\Contracts;
 
 use CraftCms\Cms\Cp\Components\ViewComponent;
-use CraftCms\Cms\Cp\Forms\Contracts\PositionableFormElement;
+use CraftCms\Cms\Cp\Forms\Contracts\FormElement;
 use CraftCms\Cms\FieldLayout\FieldLayoutFormElementContext;
 
-interface FieldLayoutFormElementProviderInterface
+interface FieldLayoutFormInputProviderInterface
 {
-    public function formElement(
+    public function inputFormElement(
         FieldLayoutFormElementContext $context,
-    ): (ViewComponent&PositionableFormElement)|null;
+    ): (ViewComponent&FormElement)|null;
 }