+++ src/FieldLayout/Exceptions/UnsupportedFieldLayoutFormElementException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CraftCms\Cms\FieldLayout\Exceptions;
+
+use LogicException;
+
+class UnsupportedFieldLayoutFormElementException extends LogicException {}